Initial commit: WIP Records status of project as at adoption of git. Thus this commit does not represent a project milestone.
Its purpose is to capture the status of the project as is, so that
we can integrate git and github workflow process into the project.
The reason for the integration of git and github is so that we can
share the project online with other interested coders..and hopefully
get contributions from them. another reason is to make use of the
fact that having a remote repository enables us not to keep copying
code on a harddisk to take home so that we can continue coding on
our home machines. Instead, we just push to the remote, then when
we get to another pc, we pull and continue working on the project.

In terms of the a vague project plan that only exists in the
author's head (lol), I envisage milestones based on a  plan like
this:-
  -complete building the UI module (this status therefore captures
   a WIP snapshot of building the UI ).
  -Build the Application module
  -Build the database server access module
  -merge the modules
  -Test and Deploy.